requirements.txt
Django==3.2.19
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
drf-yasg==1.21.7
python-dotenv==1.0.1